refactors rows to AppGridRows
